Move everything bot related to the bot package.
