+++ src/main/java/org/learn_java/bot/commands/Format.java
@@ -1,4 +1,4 @@
-package org.learn_java.commands;
+package org.learn_java.bot.commands;
 
 import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
