+++ src/main/java/org/learn_java/bot/data/repositories/InfoRepository.java
@@ -1,6 +1,6 @@
-package org.learn_java.data.repositories;
+package org.learn_java.bot.data.repositories;
 
-import org.learn_java.data.entities.Info;
+import org.learn_java.bot.data.entities.Info;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
